+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- add support for `FIXED` type
+
+### Fixed
+- Ensure `ChainSynchronizer` not scanning same blocks multiple times
 
 ## [1.2.5] - 2021-06-17
 ### Fixed